requirements.txt
simple-lama-inpainting>=0.1.2
Pillow>=9.5,<10
gradio>=5.0
rapidocr>=3.7.0
onnxruntime>=1.24.0
